+++ src/main/java/com/bitwig/extensions/controllers/novation/launchpadmini3/LpMiniHwElements.java
@@ -1,34 +1,35 @@
 package com.bitwig.extensions.controllers.novation.launchpadmini3;
 
-import com.bitwig.extension.controller.api.HardwareSurface;
-import com.bitwig.extension.controller.api.MultiStateHardwareLight;
-import com.bitwig.extensions.controllers.novation.commonsmk3.DrumButton;
-import com.bitwig.extensions.controllers.novation.commonsmk3.GridButton;
-import com.bitwig.extensions.controllers.novation.commonsmk3.LabeledButton;
-import com.bitwig.extensions.controllers.novation.commonsmk3.MidiProcessor;
-import com.bitwig.extensions.framework.di.Component;
-import com.bitwig.extensions.framework.di.Inject;
-import com.bitwig.extensions.framework.di.PostConstruct;
-
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import com.bitwig.extension.controller.api.HardwareSurface;
+import com.bitwig.extension.controller.api.MultiStateHardwareLight;
+import com.bitwig.extensions.controllers.novation.commonsmk3.DrumButton;
+import com.bitwig.extensions.controllers.novation.commonsmk3.GridButton;
+import com.bitwig.extensions.controllers.novation.commonsmk3.LabeledButton;
+import com.bitwig.extensions.controllers.novation.commonsmk3.LpHwElements;
+import com.bitwig.extensions.controllers.novation.commonsmk3.MidiProcessor;
+import com.bitwig.extensions.framework.di.Component;
+import com.bitwig.extensions.framework.di.Inject;
+import com.bitwig.extensions.framework.di.PostConstruct;
+
 @Component
-public class HwElements {
+public class LpMiniHwElements implements LpHwElements {
     private final GridButton[][] gridButtons = new GridButton[8][8];
     private final Map<LabelCcAssignmentsMini, LabeledButton> labeledButtons = new HashMap<>();
     private final List<LabeledButton> sceneLaunchButtons = new ArrayList<>();
     private final List<DrumButton> drumGridButtons = new ArrayList<>();
-
+    
     private MultiStateHardwareLight novationLight;
-
+    
     private final int[] SCENE_CCS = {0x59, 0x4f, 0x45, 0x3B, 0x31, 0x27, 0x1D, 0x13};
-
+    
     @Inject
     private MidiProcessor midiProcessor;
-
+    
     @PostConstruct
     public void init(final HardwareSurface surface) {
         initGridButtons(surface, midiProcessor);
@@ -39,14 +40,14 @@ public class HwElements {
             }
         }
         for (int i = 0; i < 8; i++) {
-            final LabeledButton sceneButton = new LabeledButton("SCENE_LAUNCH_" + (i + 1), surface, midiProcessor,
-                    SCENE_CCS[i]);
+            final LabeledButton sceneButton =
+                new LabeledButton("SCENE_LAUNCH_" + (i + 1), surface, midiProcessor, SCENE_CCS[i]);
             sceneLaunchButtons.add(sceneButton);
         }
         novationLight = surface.createMultiStateHardwareLight("NovationLight");
         novationLight.state().onUpdateHardware(hwState -> midiProcessor.updatePadLed(hwState, 0x63));
     }
-
+    
     public void refresh() {
         sceneLaunchButtons.forEach(LabeledButton::refresh);
         for (int row = 0; row < 8; row++) {
@@ -55,7 +56,7 @@ public class HwElements {
             }
         }
     }
-
+    
     private void initGridButtons(final HardwareSurface surface, final MidiProcessor midiProcessor) {
         for (int row = 0; row < 8; row++) {
             for (int col = 0; col < 8; col++) {
@@ -67,25 +68,25 @@ public class HwElements {
             drumGridButtons.add(new DrumButton(surface, midiProcessor, 8, noteValue));
         }
     }
-
+    
     public List<DrumButton> getDrumGridButtons() {
         return drumGridButtons;
     }
-
+    
     public GridButton getGridButton(final int row, final int col) {
         return gridButtons[row][col];
     }
-
+    
     public LabeledButton getLabeledButton(final LabelCcAssignmentsMini assignment) {
         return labeledButtons.get(assignment);
     }
-
+    
     public List<LabeledButton> getSceneLaunchButtons() {
         return sceneLaunchButtons;
     }
-
+    
     public MultiStateHardwareLight getNovationLight() {
         return novationLight;
     }
-
+    
 }
