+++ src/main/java/com/bitwig/extensions/controllers/novation/launchpadmini3/layers/PanSliderLayer.java
@@ -5,18 +5,18 @@ import com.bitwig.extension.controller.api.Track;
 import com.bitwig.extension.controller.api.TrackBank;
 import com.bitwig.extensions.controllers.novation.commonsmk3.MidiProcessor;
 import com.bitwig.extensions.controllers.novation.commonsmk3.SliderBinding;
-import com.bitwig.extensions.controllers.novation.launchpadmini3.HwElements;
-import com.bitwig.extensions.controllers.novation.launchpadmini3.ViewCursorControl;
+import com.bitwig.extensions.controllers.novation.commonsmk3.ViewCursorControl;
+import com.bitwig.extensions.controllers.novation.launchpadmini3.LpMiniHwElements;
 import com.bitwig.extensions.framework.Layers;
 
 public class PanSliderLayer extends TrackSliderLayer {
-
+    
     public PanSliderLayer(final ViewCursorControl viewCursorControl, final HardwareSurface controlSurface,
-                          final Layers layers, final MidiProcessor midiProcessor, final HwElements hwElements) {
+        final Layers layers, final MidiProcessor midiProcessor, final LpMiniHwElements hwElements) {
         super("PAN", controlSurface, layers, midiProcessor, 20, -1);
         bind(viewCursorControl.getTrackBank());
     }
-
+    
     @Override
     protected void bind(final TrackBank trackBank) {
         for (int i = 0; i < 8; i++) {
@@ -26,12 +26,12 @@ public class PanSliderLayer extends TrackSliderLayer {
             valueBindings.add(binding);
         }
     }
-
+    
     @Override
     protected void refreshTrackColors() {
         final boolean[] exists = trackState.getExists();
         final int[] colorIndex = trackState.getColorIndex();
-
+        
         for (int i = 0; i < 8; i++) {
             if (exists[i]) {
                 tracksExistsColors[i] = colorIndex[i];
@@ -40,7 +40,7 @@ public class PanSliderLayer extends TrackSliderLayer {
             }
         }
     }
-
+    
     @Override
     protected void updateFaderState() {
         if (isActive()) {
@@ -49,18 +49,18 @@ public class PanSliderLayer extends TrackSliderLayer {
             valueBindings.forEach(SliderBinding::update);
         }
     }
-
+    
     @Override
     protected void onActivate() {
         super.onActivate();
         refreshTrackColors();
         updateFaderState();
     }
-
+    
     @Override
     protected void onDeactivate() {
         super.onDeactivate();
         updateFaderState();
     }
-
+    
 }
