+++ src/main/java/com/bitwig/extensions/controllers/novation/launchpadmini3/layers/SessionLayer.java
@@ -22,11 +22,11 @@ import com.bitwig.extensions.controllers.novation.commonsmk3.LaunchpadDeviceConf
 import com.bitwig.extensions.controllers.novation.commonsmk3.MidiProcessor;
 import com.bitwig.extensions.controllers.novation.commonsmk3.PanelLayout;
 import com.bitwig.extensions.controllers.novation.commonsmk3.RgbState;
-import com.bitwig.extensions.controllers.novation.launchpadmini3.HwElements;
+import com.bitwig.extensions.controllers.novation.commonsmk3.ViewCursorControl;
+import com.bitwig.extensions.controllers.novation.launchpadmini3.LpMiniHwElements;
 import com.bitwig.extensions.controllers.novation.launchpadmini3.LabelCcAssignmentsMini;
 import com.bitwig.extensions.controllers.novation.launchpadmini3.LpMode;
 import com.bitwig.extensions.controllers.novation.launchpadmini3.TrackMode;
-import com.bitwig.extensions.controllers.novation.launchpadmini3.ViewCursorControl;
 import com.bitwig.extensions.controllers.novation.launchpadpromk3.FocusSlot;
 import com.bitwig.extensions.framework.Layer;
 import com.bitwig.extensions.framework.Layers;
@@ -104,7 +104,7 @@ public class SessionLayer extends AbstractLpSessionLayer {
     }
     
     @PostConstruct
-    protected void init(final ControllerHost host, final Transport transport, final HwElements hwElements) {
+    protected void init(final ControllerHost host, final Transport transport, final LpMiniHwElements hwElements) {
         clipLauncherOverdub = transport.isClipLauncherOverdubEnabled();
         clipLauncherOverdub.markInterested();
         
@@ -232,7 +232,8 @@ public class SessionLayer extends AbstractLpSessionLayer {
         }
     }
     
-    private void initNavigation(final HwElements hwElements, final TrackBank trackBank, final SceneBank sceneBank) {
+    private void initNavigation(final LpMiniHwElements hwElements, final TrackBank trackBank,
+        final SceneBank sceneBank) {
         final LabeledButton upButton = hwElements.getLabeledButton(LabelCcAssignmentsMini.UP);
         final LabeledButton downButton = hwElements.getLabeledButton(LabelCcAssignmentsMini.DOWN);
         final LabeledButton leftButton = hwElements.getLabeledButton(LabelCcAssignmentsMini.LEFT);
@@ -271,7 +272,7 @@ public class SessionLayer extends AbstractLpSessionLayer {
         
     }
     
-    private void initClipControl(final HwElements hwElements, final TrackBank trackBank) {
+    private void initClipControl(final LpMiniHwElements hwElements, final TrackBank trackBank) {
         for (int i = 0; i < 8; i++) {
             final int trackIndex = i;
             final Track track = trackBank.getItemAt(trackIndex);
@@ -295,7 +296,7 @@ public class SessionLayer extends AbstractLpSessionLayer {
         horizontalTrackControlLayer.initControlLayer(hwElements, viewCursorControl);
     }
     
-    private void initSceneControl(final HwElements hwElements, final SceneBank sceneBank) {
+    private void initSceneControl(final LpMiniHwElements hwElements, final SceneBank sceneBank) {
         sceneBank.setIndication(true);
         sceneBank.scrollPosition().addValueObserver(value -> sceneOffset = value);
         for (int i = 0; i < 8; i++) {
@@ -347,7 +348,7 @@ public class SessionLayer extends AbstractLpSessionLayer {
         return colorIndex;
     }
     
-    private void initTrackControlXSceneButtons(final HwElements hwElements, final Layer layer) {
+    private void initTrackControlXSceneButtons(final LpMiniHwElements hwElements, final Layer layer) {
         initVolumeControl(hwElements, layer, 0);
         initPanControl(hwElements, layer, 1);
         initSendsAControl(hwElements, layer, 2);
@@ -358,7 +359,7 @@ public class SessionLayer extends AbstractLpSessionLayer {
         initArmControl(hwElements, layer, 7);
     }
     
-    private void initTrackControlSceneButtons(final HwElements hwElements, final Layer layer) {
+    private void initTrackControlSceneButtons(final LpMiniHwElements hwElements, final Layer layer) {
         initVolumeControl(hwElements, layer, 0);
         initPanControl(hwElements, layer, 1);
         initSendsAControl(hwElements, layer, 2);
@@ -369,7 +370,7 @@ public class SessionLayer extends AbstractLpSessionLayer {
         initSoloControl(hwElements, layer, 7);
     }
     
-    private void initVolumeControl(final HwElements hwElements, final Layer layer, final int index) {
+    private void initVolumeControl(final LpMiniHwElements hwElements, final Layer layer, final int index) {
         final LabeledButton volumeButton = hwElements.getSceneLaunchButtons().get(index);
         volumeButton.bindPressReleaseAfter(this, () -> intoControlMode(ControlMode.VOLUME), this::returnToPreviousMode,
             MOMENTARY_TIME);
@@ -377,7 +378,7 @@ public class SessionLayer extends AbstractLpSessionLayer {
             () -> controlMode == ControlMode.VOLUME ? RgbState.of(9) : RgbState.of(MODE_INACTIVE_COLOR));
     }
     
-    private void initPanControl(final HwElements hwElements, final Layer layer, final int index) {
+    private void initPanControl(final LpMiniHwElements hwElements, final Layer layer, final int index) {
         final LabeledButton panButton = hwElements.getSceneLaunchButtons().get(index);
         panButton.bindPressReleaseAfter(this, () -> intoControlMode(ControlMode.PAN), this::returnToPreviousMode,
             MOMENTARY_TIME);
@@ -385,21 +386,21 @@ public class SessionLayer extends AbstractLpSessionLayer {
             () -> controlMode == ControlMode.PAN ? RgbState.of(9) : RgbState.of(MODE_INACTIVE_COLOR));
     }
     
-    private void initSendsAControl(final HwElements hwElements, final Layer layer, final int index) {
+    private void initSendsAControl(final LpMiniHwElements hwElements, final Layer layer, final int index) {
         final LabeledButton sendsAButton = hwElements.getSceneLaunchButtons().get(index);
         sendsAButton.bindPressReleaseAfter(this, () -> intoControlMode(ControlMode.SENDS_A), this::returnToPreviousMode,
             MOMENTARY_TIME);
         sendsAButton.bindLight(layer, () -> getSendsState(ControlMode.SENDS_A));
     }
     
-    private void initSendsBControl(final HwElements hwElements, final Layer layer, final int index) {
+    private void initSendsBControl(final LpMiniHwElements hwElements, final Layer layer, final int index) {
         final LabeledButton sendsBButton = hwElements.getSceneLaunchButtons().get(index);
         sendsBButton.bindPressReleaseAfter(this, () -> intoControlMode(ControlMode.SENDS_B), this::returnToPreviousMode,
             MOMENTARY_TIME);
         sendsBButton.bindLight(layer, () -> getSendsState(ControlMode.SENDS_B));
     }
     
-    private void initDeviceControl(final HwElements hwElements, final Layer layer, final int index) {
+    private void initDeviceControl(final LpMiniHwElements hwElements, final Layer layer, final int index) {
         final LabeledButton deviceButton = hwElements.getSceneLaunchButtons().get(index);
         deviceButton.bindPressReleaseAfter(this, () -> intoControlMode(ControlMode.DEVICE), this::returnToPreviousMode,
             MOMENTARY_TIME);
@@ -407,7 +408,7 @@ public class SessionLayer extends AbstractLpSessionLayer {
             () -> controlMode == ControlMode.DEVICE ? RgbState.of(33) : RgbState.of(MODE_INACTIVE_COLOR));
     }
     
-    private void initStopControl(final HwElements hwElements, final Layer layer, final int index) {
+    private void initStopControl(final LpMiniHwElements hwElements, final Layer layer, final int index) {
         final LabeledButton stopButton = hwElements.getSceneLaunchButtons().get(index);
         stopButton.bindPressReleaseAfter(this, () -> intoTrackMode(TrackMode.STOP), this::returnToPreviousMode,
             MOMENTARY_TIME);
@@ -415,7 +416,7 @@ public class SessionLayer extends AbstractLpSessionLayer {
             () -> trackMode == TrackMode.STOP ? RgbState.of(5) : RgbState.of(MODE_INACTIVE_COLOR));
     }
     
-    private void initMuteControl(final HwElements hwElements, final Layer layer, final int index) {
+    private void initMuteControl(final LpMiniHwElements hwElements, final Layer layer, final int index) {
         final LabeledButton muteButton = hwElements.getSceneLaunchButtons().get(index);
         muteButton.bindPressReleaseAfter(this, () -> intoTrackMode(TrackMode.MUTE), this::returnToPreviousMode,
             MOMENTARY_TIME);
@@ -423,7 +424,7 @@ public class SessionLayer extends AbstractLpSessionLayer {
             () -> trackMode == TrackMode.MUTE ? RgbState.of(9) : RgbState.of(MODE_INACTIVE_COLOR));
     }
     
-    private void initSoloControl(final HwElements hwElements, final Layer layer, final int index) {
+    private void initSoloControl(final LpMiniHwElements hwElements, final Layer layer, final int index) {
         final LabeledButton soloButton = hwElements.getSceneLaunchButtons().get(index);
         soloButton.bindPressReleaseAfter(this, () -> intoTrackMode(TrackMode.SOLO), this::returnToPreviousMode,
             MOMENTARY_TIME);
@@ -431,7 +432,7 @@ public class SessionLayer extends AbstractLpSessionLayer {
             () -> trackMode == TrackMode.SOLO ? RgbState.of(13) : RgbState.of(MODE_INACTIVE_COLOR));
     }
     
-    private void initArmControl(final HwElements hwElements, final Layer layer, final int index) {
+    private void initArmControl(final LpMiniHwElements hwElements, final Layer layer, final int index) {
         final LabeledButton soloButton = hwElements.getSceneLaunchButtons().get(index);
         soloButton.bindPressReleaseAfter(this, () -> intoTrackMode(TrackMode.ARM), this::returnToPreviousMode,
             MOMENTARY_TIME);
