+++ src/main/java/com/bitwig/extensions/controllers/novation/launchpadpromk3/sliderlayers/DeviceSliderLayer.java
@@ -1,18 +1,27 @@
 package com.bitwig.extensions.controllers.novation.launchpadpromk3.sliderlayers;
 
-import com.bitwig.extension.controller.api.*;
+import com.bitwig.extension.controller.api.CursorDeviceLayer;
+import com.bitwig.extension.controller.api.CursorRemoteControlsPage;
+import com.bitwig.extension.controller.api.DeviceBank;
+import com.bitwig.extension.controller.api.HardwareSurface;
+import com.bitwig.extension.controller.api.PinnableCursorDevice;
+import com.bitwig.extension.controller.api.RemoteControl;
 import com.bitwig.extensions.controllers.novation.commonsmk3.LabeledButton;
 import com.bitwig.extensions.controllers.novation.commonsmk3.MidiProcessor;
 import com.bitwig.extensions.controllers.novation.commonsmk3.RgbState;
 import com.bitwig.extensions.controllers.novation.commonsmk3.SliderBinding;
-import com.bitwig.extensions.controllers.novation.launchpadpromk3.*;
+import com.bitwig.extensions.controllers.novation.commonsmk3.ViewCursorControl;
+import com.bitwig.extensions.controllers.novation.launchpadpromk3.LpProHwElements;
+import com.bitwig.extensions.controllers.novation.launchpadpromk3.LabelCcAssignments;
+import com.bitwig.extensions.controllers.novation.launchpadpromk3.LpBaseMode;
+import com.bitwig.extensions.controllers.novation.launchpadpromk3.SysExHandler;
 import com.bitwig.extensions.controllers.novation.launchpadpromk3.layers.ControlMode;
 import com.bitwig.extensions.framework.Layers;
 
 public class DeviceSliderLayer extends SliderLayer {
-
+    
     private static final int[] PARAM_COLORS = {5, 9, 13, 25, 29, 41, 49, 57};
-
+    
     private final CursorRemoteControlsPage parameterBank;
     private final DeviceBank drumDeviceBank;
     final PinnableCursorDevice device;
@@ -23,10 +32,10 @@ public class DeviceSliderLayer extends SliderLayer {
     private boolean hasLayers;
     private boolean hasSlots;
     private String[] deviceSlotNames = new String[0];
-
+    
     public DeviceSliderLayer(final ViewCursorControl viewCursorControl, final HardwareSurface controlSurface,
-                             final Layers layers, final MidiProcessor midiProcessor, final SysExHandler sysExHandler,
-                             final HwElements hwElements) {
+        final Layers layers, final MidiProcessor midiProcessor, final SysExHandler sysExHandler,
+        final LpProHwElements hwElements) {
         super("DEVICE", ControlMode.DEVICE, controlSurface, layers, midiProcessor, sysExHandler);
         device = viewCursorControl.getCursorDevice();
         parameterBank = device.createCursorRemoteControlsPage(8);
@@ -35,53 +44,53 @@ public class DeviceSliderLayer extends SliderLayer {
         final PinnableCursorDevice primary = viewCursorControl.getPrimaryDevice();
         final CursorDeviceLayer drumCursor = primary.createCursorLayer();
         drumDeviceBank = drumCursor.createDeviceBank(8);
-
+        
         initSceneButtons(hwElements);
         for (int i = 0; i < 8; i++) {
             final RemoteControl parameter = parameterBank.getParameter(i);
-
+            
             final SliderBinding binding = new SliderBinding(mode.getCcNr(), parameter, sliders[i], i, midiProcessor);
             addBinding(binding);
             valueBindings.add(binding);
         }
         initNavigation(hwElements);
     }
-
-    private void initNavigation(final HwElements hwElements) {
+    
+    private void initNavigation(final LpProHwElements hwElements) {
         final LabeledButton upButton = hwElements.getLabeledButton(LabelCcAssignments.UP);
         final LabeledButton downButton = hwElements.getLabeledButton(LabelCcAssignments.DOWN);
         final LabeledButton leftButton = hwElements.getLabeledButton(LabelCcAssignments.LEFT);
         final LabeledButton rightButton = hwElements.getLabeledButton(LabelCcAssignments.RIGHT);
         parameterBank.hasNext().markInterested();
         parameterBank.hasPrevious().markInterested();
-
+        
         device.hasNext().markInterested();
         device.hasPrevious().markInterested();
-
+        
         device.isNested().addValueObserver(nested -> isNested = nested);
         device.slotNames().addValueObserver(slotNames -> deviceSlotNames = slotNames);
         device.hasSlots().addValueObserver(hasSlots -> this.hasSlots = hasSlots);
         device.hasLayers().addValueObserver(hasLayers -> this.hasLayers = hasLayers);
         device.hasDrumPads().addValueObserver(hasPads -> hasDrumPads = hasPads);
-
+        
         final RgbState scrollColor = RgbState.of(23);
         upButton.bindRepeatHold(this, () -> handleNavigateDevice(-1));
         upButton.bindLight(this, () -> isNested ? scrollColor : RgbState.OFF);
-
+        
         downButton.bindRepeatHold(this, () -> handleNavigateDevice(1));
         downButton.bindLight(this, () -> canNavigateDown() ? scrollColor : RgbState.OFF);
-
+        
         leftButton.bindRepeatHold(this, () -> handleNavigateChain(-1));
         leftButton.bindLight(this, () -> device.hasPrevious().get() ? scrollColor : RgbState.OFF);
-
+        
         rightButton.bindRepeatHold(this, () -> handleNavigateChain(1));
         rightButton.bindLight(this, () -> device.hasNext().get() ? scrollColor : RgbState.OFF);
     }
-
+    
     private boolean canNavigateDown() {
         return hasDrumPads || hasLayers || hasSlots;
     }
-
+    
     private void handleNavigateChain(final int amount) {
         if (amount > 0) {
             device.selectNext();
@@ -89,7 +98,7 @@ public class DeviceSliderLayer extends SliderLayer {
             device.selectPrevious();
         }
     }
-
+    
     private void handleNavigateDevice(final int amount) {
         if (amount > 0) {
             if (device.hasDrumPads().get()) {
@@ -103,7 +112,7 @@ public class DeviceSliderLayer extends SliderLayer {
             device.selectParent();
         }
     }
-
+    
     private void handleNavigateParameters(final int amount, final CursorRemoteControlsPage parameters) {
         if (amount > 0) {
             parameters.selectNextPage(false);
@@ -111,9 +120,9 @@ public class DeviceSliderLayer extends SliderLayer {
             parameters.selectPreviousPage(false);
         }
     }
-
-
-    private void initSceneButtons(final HwElements hwElements) {
+    
+    
+    private void initSceneButtons(final LpProHwElements hwElements) {
         for (int i = 0; i < 8; i++) {
             final int index = i;
             final LabeledButton sceneButton = hwElements.getSceneLaunchButtons().get(index);
@@ -121,11 +130,11 @@ public class DeviceSliderLayer extends SliderLayer {
             sceneButton.bindLight(this, () -> getColor(index));
         }
     }
-
+    
     private void handleSendSelect(final int index) {
         parameterBank.selectedPageIndex().set(index);
     }
-
+    
     private RgbState getColor(final int index) {
         if (index < parameterPages) {
             if (pageIndex == index) {
@@ -135,12 +144,12 @@ public class DeviceSliderLayer extends SliderLayer {
         }
         return RgbState.OFF;
     }
-
+    
     @Override
     protected void refreshTrackColors() {
         System.arraycopy(PARAM_COLORS, 0, tracksExistsColors, 0, PARAM_COLORS.length);
     }
-
+    
     @Override
     protected void onActivate() {
         super.onActivate();
@@ -148,5 +157,5 @@ public class DeviceSliderLayer extends SliderLayer {
         modeHandler.setFaderBank(0, mode, tracksExistsColors);
         modeHandler.changeMode(LpBaseMode.FADER, mode.getBankId());
     }
-
+    
 }
