+++ src/main/java/com/bitwig/extensions/controllers/novation/launchpadpromk3/sliderlayers/PanSliderLayer.java
@@ -6,26 +6,26 @@ import com.bitwig.extension.controller.api.TrackBank;
 import com.bitwig.extensions.controllers.novation.commonsmk3.LabeledButton;
 import com.bitwig.extensions.controllers.novation.commonsmk3.MidiProcessor;
 import com.bitwig.extensions.controllers.novation.commonsmk3.SliderBinding;
-import com.bitwig.extensions.controllers.novation.launchpadpromk3.HwElements;
+import com.bitwig.extensions.controllers.novation.commonsmk3.ViewCursorControl;
+import com.bitwig.extensions.controllers.novation.launchpadpromk3.LpProHwElements;
 import com.bitwig.extensions.controllers.novation.launchpadpromk3.LpBaseMode;
 import com.bitwig.extensions.controllers.novation.launchpadpromk3.SysExHandler;
-import com.bitwig.extensions.controllers.novation.launchpadpromk3.ViewCursorControl;
 import com.bitwig.extensions.controllers.novation.launchpadpromk3.layers.ControlMode;
 import com.bitwig.extensions.framework.Layer;
 import com.bitwig.extensions.framework.Layers;
 
 public class PanSliderLayer extends TrackSliderLayer {
     private final Layer sceneButtonLayer;
-
+    
     public PanSliderLayer(final ViewCursorControl viewCursorControl, final HardwareSurface controlSurface,
-                          final Layers layers, final MidiProcessor midiProcessor, final SysExHandler sysExHandler,
-                          final HwElements hwElements) {
+        final Layers layers, final MidiProcessor midiProcessor, final SysExHandler sysExHandler,
+        final LpProHwElements hwElements) {
         super("PAN", ControlMode.PAN, controlSurface, layers, midiProcessor, sysExHandler);
         bind(viewCursorControl.getTrackBank());
         sceneButtonLayer = new Layer(layers, "VOL_SCENE_BUTTONS");
         initSceneButtons(hwElements);
     }
-
+    
     @Override
     protected void bind(final TrackBank trackBank) {
         for (int i = 0; i < 8; i++) {
@@ -35,12 +35,12 @@ public class PanSliderLayer extends TrackSliderLayer {
             valueBindings.add(binding);
         }
     }
-
+    
     @Override
     protected void refreshTrackColors() {
         final boolean[] exists = trackState.getExists();
         final int[] colorIndex = trackState.getColorIndex();
-
+        
         for (int i = 0; i < 8; i++) {
             if (exists[i]) {
                 tracksExistsColors[i] = colorIndex[i];
@@ -49,15 +49,15 @@ public class PanSliderLayer extends TrackSliderLayer {
             }
         }
     }
-
-    private void initSceneButtons(final HwElements hwElements) {
+    
+    private void initSceneButtons(final LpProHwElements hwElements) {
         for (int index = 0; index < 8; index++) {
             final LabeledButton sceneButton = hwElements.getSceneLaunchButtons().get(index);
             sceneButton.disable(sceneButtonLayer);
         }
     }
-
-
+    
+    
     @Override
     protected void onActivate() {
         super.onActivate();
@@ -66,13 +66,13 @@ public class PanSliderLayer extends TrackSliderLayer {
         modeHandler.setFaderBank(1, mode, tracksExistsColors);
         modeHandler.changeMode(LpBaseMode.FADER, mode.getBankId());
     }
-
-
+    
+    
     @Override
     protected void onDeactivate() {
         super.onDeactivate();
         sceneButtonLayer.setIsActive(false);
         updateFaderState();
     }
-
+    
 }
