+++ src/main/java/com/bitwig/extensions/framework/di/PackageHelper.java
@@ -4,7 +4,11 @@ import java.io.File;
 import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Enumeration;
+import java.util.List;
+import java.util.Optional;
+import java.util.TreeSet;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 
@@ -12,17 +16,27 @@ import java.util.zip.ZipInputStream;
  * Utility class to get classes from packages.
  */
 public class PackageHelper {
-
+    
     private PackageHelper() {
         // just a  utility class
     }
-
-    public static List<Class<?>> getClasses(final Class<?> baseClass) throws IOException, ClassNotFoundException {
-        return getClasses(baseClass.getPackageName(), baseClass.getClassLoader());
+    
+    public static List<Class<?>> getClasses(final Class<?> baseClass, final Package... packages) throws
+        IOException,
+        ClassNotFoundException {
+        final ClassLoader classLoader = baseClass.getClassLoader();
+        final List<Class<?>> classes = new ArrayList<>();
+        final List<Class<?>> baseList = getClasses(baseClass.getPackageName(), classLoader);
+        for (final Package pack : packages) {
+            classes.addAll(getClasses(pack.getName(), classLoader));
+        }
+        classes.addAll(baseList);
+        return classes;
     }
-
-    public static List<Class<?>> getClasses(final String packageName,
-                                            final ClassLoader classLoader) throws IOException, ClassNotFoundException {
+    
+    private static List<Class<?>> getClasses(final String packageName, final ClassLoader classLoader) throws
+        IOException,
+        ClassNotFoundException {
         assert classLoader != null;
         final String path = packageName.replace('.', '/');
         final Enumeration<URL> resources = classLoader.getResources(path);
@@ -31,7 +45,7 @@ public class PackageHelper {
             final URL resource = resources.nextElement();
             dirs.add(resource.getFile());
         }
-
+        
         final TreeSet<String> classes = new TreeSet<>();
         for (final String directory : dirs) {
             classes.addAll(findClasses(directory, packageName));
@@ -40,11 +54,11 @@ public class PackageHelper {
         for (final String clazz : classes) {
             classList.add(Class.forName(clazz));
         }
-
+        
         return classList;
     }
-
-
+    
+    
     private static Optional<URL> toDirectoryFilePath(final String directory) throws MalformedURLException {
         if (directory.startsWith("file:") && directory.contains("!")) {
             final String[] split = directory.split("!");
@@ -52,41 +66,39 @@ public class PackageHelper {
         }
         return Optional.empty();
     }
-
+    
     private static String toSystemPath(final String directory) {
         if (File.separatorChar == '\\') {
             return directory.replace("%20", " ");
         }
         return directory;
     }
-
+    
     private static Optional<String> classNameFromZipEntry(final ZipEntry entry, final String packageName) {
         if (entry.getName().endsWith(".class")) {
-            final String className = entry.getName()
-                    .replaceAll("[$].*", "")
-                    .replaceAll("[.]class", "")
-                    .replace('/', '.');
+            final String className =
+                entry.getName().replaceAll("[$].*", "").replaceAll("[.]class", "").replace('/', '.');
             if (className.startsWith(packageName)) {
                 return Optional.of(className);
             }
         }
         return Optional.empty();
     }
-
+    
     private static TreeSet<String> findClasses(final String directory, final String packageName) throws IOException {
         final TreeSet<String> classes = new TreeSet<>();
-
+        
         final Optional<URL> dirUrl = toDirectoryFilePath(directory);
         if (dirUrl.isPresent()) {
             try (final ZipInputStream zip = new ZipInputStream(dirUrl.get().openStream())) {
                 ZipEntry entry;
                 while ((entry = zip.getNextEntry()) != null) {
                     classNameFromZipEntry(entry, packageName) //
-                            .ifPresent(classes::add);
+                        .ifPresent(classes::add);
                 }
             }
         }
-
+        
         final File dir = new File(toSystemPath(directory));
         if (!dir.exists()) {
             return classes;
